\name{ex7}
\alias{myP}
\title{Calculte probabilities of continuous random variables}
\usage{
  myP(g, opt)
}
\description{
  Crearea unei funcții P care permite calculul diferitelor tipuri de probabilități asociate unei variabile aleatoare continue (similar funcției P din pachetul discreteRV)
}
\examples{
  myP(g,"x>1|x<1.5")
}
