\name{ex8}
\alias{Synthesis sheet}
\title{Synthesis sheet}
\usage{
  output_details(name)
}
\description{
  Afișarea unei “fișe de sinteză” care să conțină informații de bază despre respectiva repartiție (cu precizarea sursei informației!). Relevant aici ar fi să precizați pentru ce e folosită ȋn mod uzual acea repartiție, semnificația parametrilor, media, dispersia etc.
}
\examples{
  output_details("uniform")
  output_details("exponential")
  output_details("normal")
  output_details("pareto")
  output_details("cauchy")
  output_details("logistic")
  output_details("weibull")
}
